Update script.js
Caching API Responses:
Cache the responses of the API for cities that have already been requested. To the load on the server and speed up subsequent requests.

Batch API Calls:
This reduces the number of Network requests and improves performance.

Minimize Data Payload:
Modified the API query to request only essential data and not the entire dataset.